+++ app/containers/OrderPage/OrderDetailPage/OrderDetail/constants.tsx
@@ -2,12 +2,6 @@ import * as React from 'react';
 
 import { Col, Row, Steps, List } from 'antd';
 
-export const practiceDetailData = [
-  ['Name:', 'Practice 4399'],
-  ['Phone:', '(505) 761-5555'],
-  ['Practice:', 'Partner Practice'],
-  ['DVM ID:', '4399'],
-];
 export const clientDetailData = [
   ['Name:', 'Nikko Cabrera'],
   ['Phone 1:', '(623) 521-1111'],
