+++ app/containers/OrderPage/OrderDetailPage/OrderDetail/index.tsx
@@ -6,17 +6,22 @@
  */
 
 // remove, use babel for antD css
-import { Layout } from 'antd';
+import { Layout, Row } from 'antd';
 import 'antd/dist/antd.css';
 
 import * as React from 'react';
 import OrderDetailHeader from '../OrderDetailHeader';
+import PracticeInformation from '../OrderDetailPractice';
+
 
 export default function OrderDetail() {
 
   return (
     <Layout>
       <OrderDetailHeader/>
+        <Row gutter={16} align="top" style={{ marginTop: '20px' }}>
+          <PracticeInformation/>
+        </Row>
     </Layout>
   );
 }
