+++ app/containers/OrderPage/OrderDetailPage/OrderDetailPractice/constants.tsx
@@ -1,0 +1,6 @@
+export const practiceDetailData = [
+  ['Name:', 'Practice 4399'],
+  ['Phone:', '(505) 761-5555'],
+  ['Practice:', 'Partner Practice'],
+  ['DVM ID:', '4399'],
+];
